% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_axes1.r
\name{alt_axes}
\alias{alt_axes}
\title{Alternates ticks so they do not crowd each other}
\usage{
alt_axes(i)
}
\arguments{
\item{i}{An integer value. a subplot counter, from 1 to the total number of subplots = prod(par("mfrow"))}
}
\description{
These lines provide a verbose description.
This could be multiple lines
}
\examples{
# Some code

}
